Allow markup and links in the info bar.
This changes the behavior of
anaconda_base_window_set_{error,info_warning}, in that the message
parameter is now treated as markup. Encourage the use of links in the
message in order to add something activatable to the tab order. The
actual link URI will be ignored.

Change the info-bar-clicked event to activate on button-pressed instead
of button-released, which keeps us from receiving both a button event
and a link event if someone clicks on a link.

(cherry picked from commit 873f816bce6f4ed4611dd98765ba276ed56279ee)

Related: rhbz#1125145