Display an error for exceptions during GUI setup
Exceptions can occur between the start of GUI initialization and the Gtk
main loop, in particular if product images specify missing or bad
stylesheets. If an exception occurs in this code, the GUI must release
the GUI lock so that python-meh knows to start its own main loop.
Otherwise meh will enqueue itself in a main loop that will never be run,
and anaconda exits and reboots the system.