Add a check for invisible_char validity
The invisible_char setting is usually non-ASCII, so sometimes, say if
someone opens a glade file in a non-unicode-aware editor and then saves
the file, the string gets trashed to ? or ???. Check that this didn't
happen.

Also add a check that invisible_char is needed at all since it's just
glade noise more often than not.