Keyboard Scroll Fails in Viewports (#1189899)
Robert Marshall's cherry pick of David Shea's original commit-id:

35324b3fd891400c0405c50a18798080910627ed

This patch corrects an issue where users tried to navigate through
the storage or software selections with a keyboard and the window
failed to scroll when the selection was not in the current viewport.

It does not include changes to filter.py and filter.glade per
consultation with David Shea.

Resolves: rhbz#1189899