Connect viewport adjustments to child focus adjustments (#1192155)
When we use GtkScrolledWindow with non-natively scrollable widgets, the
scroll adjustments on the intermediate viewport need to be connected to
the focus adjustments on non-scrollable container in order for the
viewport to automatically scroll on keyboard navigation events. Normally
this is done automatically by gtk, but adding silent surprises in the
middle of the widget hierarchy breaks glade so glade requires the
viewport be explicit, and glade does not allow one to change any of the
adjustment settings we need to get to. See
https://bugzilla.gnome.org/show_bug.cgi?id=744721

I am not optimistic about glade changing things quickly or ever so let's
just work around it and have things work a little nicer.

(cherry picked from commit 35324b3fd891400c0405c50a18798080910627ed)