Add rc-release target
Use these make targets for creating scratch rc builds that can be used
before an official release. rc-release will run scratch-bumpver to bump
the version and add a rc release with the current date and time to the
specfile.  scratch is then run to create the tar.gz file and then mock
is used to build a rc rpm package.

No changes are committed to git, no zanata account is required and
nothing is pushed to zanata.

This also Modifes scratch-bumpver to now include the rc release number.

This requires that mock is installed and the current user is a member of
the mock group.

Build the rc like this:
git clean -xdf
git reset --hard
./autogen.sh
./configure --disable-static --enable-introspection --enable-gtk-doc
make rc-release

If everything goes well there will be a pile of rpms in the current
directory.