Handle /boot on btrfs for live (#1200539)
live has problems with /boot on btrfs because there is no initial
initramfs so grub.cfg is created without an initrd line and grubby
doesn't add it when the initramfs is generated.

This kludge runs writeBootLoader a second time *only* if this is a live
install with /boot on btrfs.