Generate the pyanaconda module documentation
readthedocs.org can't install pyanaconda (no setup.py) so generate the
modules.rst and pyanaconda.*rst documentation oursevles. This isn't ideal,
but is better than no module documentation.

The modules.rst and pyanaconda*rst files should not be edited, they are
generated by make apidoc in the docs/ directory.

This also adds pulling the documentation version number from
configure.ac, and rebuilding the apidoc files after bumpver and
scratch-bumpver so that the pyanaconda module is updated with each
build.

Note that documentation updates require that sphinx be installed, the
apidoc target will skip updating the files, leaving the existing
versions in place if /usr/bin/sphinx-apidoc is not present.