Allow overriding password policy check
The Fedora Engineering Steering Committee decided that for Fedora
22, we would restore the double-click of "done" to override the
password policy check. This is a temporary measure while we work
on a distro-wide security policy to be implemented in Fedora 23.

At present, all three passwords (root, user and luks) will retain
the same policy: minimum length of 6, minimum pwquality score of
50 and the ability to override this with a double-click of done.

Details:
https://fedorahosted.org/fesco/ticket/1412#comment:40
https://fedoraproject.org/wiki/User:Kevin/Draft_Passwordpolicy