Add --tmpdir to parse-kickstart for testing
This will allow the /tmp directory to be overridden for testing
purposes. Pass --tmpdir /path/to/tmp/ to override it. Note that the =
syntax is *not* supported, they must be separate arguments.

This also changes the ifcfg files to have their keys sorted so that
testing the output is easier.