Preserve the order of boot args added by kickstart.
This adds OrderedSet to anaconda, just like it is in pykickstart.  I have added
additional functionality that we need to the Arguments class in bootloader.py
so if OrderedSet is needed elsewhere, it won't inherit the special behavior.

This also attempts to keep "rhgb quiet" at the end of the line by moving where
passed-in command line arguments are processed to the very end.  Note that the
storage args get popped off much earlier.

Resolves: rhbz#1188948