Make StorageSpoke's on_back_clicked less complicated
This method has grown into something very long, complicated and hard to
understand as a whole. It happened many times in the past that when doing
changes we forgot to cover one of the many branches its code code could go
through.

In order to mitigate the risk of such errors in the future this patch moves some
parts of the code into separate methods, names some contants and deduplicates
some checks previously done multiple times with zero risk of changes happening
in the meantime. Also it unifies the response codes of the dialogs by changing
the ResizeDialog's "Reclaim" button's response ID to 1 (meaning OK).