Create and use snapshot of on-disk storage with no modifications (#1166598)
Reverting changes in storage is really complicated and trying to do so lead us
to many hard bugs like the one referenced above. So instead of playing this
tricky game, let's just create a snapshot of unmodified on-disk storage and
revert everything to it whenever we need to start over.