Add support for chrony pool directive
With chrony-2.0 it is possible to specify a pool of NTP servers instead
of individual servers. It selects four servers from the addresses that
the pool name is resolved to and replaces them automatically when they
are unreachable or falsetickers. This is now used in the default Fedora
chrony configuration.

To avoid extending the ntp API and the date time spoke, pools are simply
presented as four separate servers with the same name.

Resolves: rhbz#1193389