Fix issues with advanced storage searching.
Search by Port / Target / LUN was usually returning an empty list since
the port test was failing to compare a string to an int. Beyond that, do
not return entries with no port when a port is selected, and do not
return iscsi entries when a non-integer LUN is entered.

Add an empty entry to all of the filter combo boxes so that the combo
box criterion can be cleared.

Replace a builder.get_object call with the already gotten object
reference.

(cherry picked from commit e13a970648b88c9b25b70763dd39478e59aa0355)

Resolves: rhbz#1168950