Check for images/install.img first for netboot
boot.iso is now built with the stage2 in images/install.img instead of
LiveOS/squashfs.img -- netboot needs to check for install.img first,
then fall back to squashfs.img if it isn't found.

This also adds a warning if neither can be retrieved, including the
interface it was using. We can't just enter the emergency shell when
this fails because other nics may become active and succeed in
retrieving it.