Fix some minor inconsistencies in filter.glade
Set no_show_all=True, can_focus=False, and show_border=False for each of
the search type notebooks. Set headers_clickable=False on the
otherTreeView.

Replace the stock gtk-find button on the other tab with a regularly
labelled button. Add a translation context to the refresh button.

Set a spacing of 5 in one of the boxes to 6.

Fix the mnemonic_widget for the multipath filter label. Add a
mnemonic_widget to the zFCP filter label.

Remove an extra er from otherIDRenderer.