Distinguish between NTP pools and servers in GUI
While we do some dirty tricks internally to support both NTP pools and servers
in kickstart (memory representation), we should just let users work with both
categories natively and do the dirty work in background.

This adds checkboxes to the NTP server configuration dialog that can be used to
mark the given URL as NTP pool. Also it changes magical column numbers to
constants as the number of columns has grown to four and definitely become
confusing.

Related: rhbz#1193389