Set the GUI-selected environment in the ksdata (#1192100)
This way an environment selected in the GUI and an environment selected
in kickstart are treated in the same way. This fixes the problem where
changing payloads forgets the previous environment selection, and it
changes the generated %packages section of anaconda-ks.cfg to contain
the environment information instead of a list of the groups in the
environment.

PackagePayload.selectEnvironment now only sets the environment in the
ksdata and does not modify the group lists. The PackagePayload
implementations must instead apply the ksdata environment to the
internal package selections when ksdata selections are applied.