Force an encoding of utf-8 on liveinst installs (#1217504)
This adds a copy of sitecustomize.py to a directory that can be added
to PYTHONPATH, and adds that directory to PYTHONPATH when running
liveinst. This way anaconda do perform early sys changes while not
affecting python globally and not messing up regular imports too badly.

This does not fix unicode problems for all cases (image/dir install are
still affected), but it papers over them for live installs.