Wrap device labels earlier (#1212586)
A minimum width of 35 in the device label causes problems with smaller
resolutions (800x600). Remove width_chars so that the label wraps
earlier than the max width if it needs to, and change the wrap mode from
word to word-char to ensure that device names with extra long words
get wrapped ugly instead of causing layout issues.