iSCSI Name Validation using regexes
When adding iSCSI target disk during installation, the iSCSI Initiator name and
the Target IP is validated to enable/disable the "Start Discovery" button on the
page.  Currently, the validation of iSCSI Initiator Name is based on the
existence of dots only (".") and this can be more efficiently done by using a
regular expression.

Signed-off-by: Sujith Pandel <sujithpshankar@gmail.com>
Signed-off-by: Vratislav Podzimek <vpodzime@redhat.com>