Fix Issue with Spaces in Network IPs and Bad Input(#1174451)
This Robert Marshall's cherry pick of Jiri Konecny's original
patches to master with the following commit IDs:

9371221734c57430829056ac5fc38dedaeeff727
d2e7ea401f86dbe8717cd169dc0051c934c6fcfd

The goal of his patch was to fix a bad warning message, however, it
also as a side effect resolves an issue where placing a space into
a nameserver IP address would cause the installer to crash. This
patch also verifies IPv4 and IPv6 addresses in general.

The cherry pick removed the pieces of the original patch that
applied to the password policy present code in the current
master that are not in the RHEL 7 codebase.

Resolves: rhbz#1174451