Don't run run_report.sh from within run_kickstart_tests.sh.
Running it like this means the return value from the big loop is always the
return value of tee, which is very likely to be 0.  Thus it will look like
tests are always succeeding.  The proper place to run this script is either
by hand against run_kickstart_tests.log, or from whatever script calls
run_kickstart_tests.sh.