Disable Label Entry for PPC PReP Boot (#1172773)
The manual partitioning window in the installer allowed users to
type in and attempt to set a label on PPC PReP Boot filesystems.
This would cause the installer to register an error and go back to the
previous settings. Updated the behavior to automatically clear the
label if the PPC PReP Boot filesystem type is chosen and disabled
the label entry field to prevent users from being able to enter
information that is invalid for this filesystem type. This fix will
also do the same for any other filesystem format that does not
support labels.

Resolves: rhbz#1172773