Switch from testing for emacs* to kacst*.
The former wants to install a giant pile of stuff that takes forever (and
probably a lot of space too) which leads to the test timing out.  The latter
is a handful of font packages that do not have much in the way of external
dependencies, so should take a lot less time.