Remove yumpayload.py, its support files, and most references to yum.
The nodnf option is also gone, because it doesn't do anything anymore.  The
yum versions of kickstart_tests are no longer needed.  In a lot of places
where we reference yum (comments, suggesting "yum install whatever", etc.),
change those to reference to dnf.  A couple of those cannot change due to
being directories or log files people have come to expect.

Note I've also added a couple FIXMEs in places where I do not know if
something is still necessary or not.