Allow passing kickstart tests to be run on the command line.
This makes it way easier to just run a single test - you simply provide
it as a command line argument.  If no tests are given as arguments, it
defaults to using find as before.

This also involved reorganizing command line processing a bit to also let it
accept the same command line arguments as run_one_ks.sh does.