Add a method to execute timed actions early
Add a method to TimedAction to cancel the timer and execute the function
early, and a property to determine whether a timer is currently active.

Rearrange the guts of the decorator so that the TimedAction object is
accessible as the decorated function (mostly), and is callable as the
decorated function. This removes the need for the multi-layered inner
functions and replaces them with opaque python magic.