dd_extract: -l should not extract modules+firmware
The -l|--libraries flag to dd_extract is (apparently) supposed to only
unpack the contents of /lib, /lib64, /usr/lib, and /usr/lib64.

Except... kernel modules are in /lib/modules, and firmware is in
/lib/firmware. So --libraries also extracts firmware and modules,
if present.

This patch makes it exclude files under /lib/modules and /lib/firmware
when using --libraries.