Rearrange filter.glade the way glade wants it now
This alphabetizes the properties, removes some properties that are equal
to the default value, adds some empty placeholders, and adds the
internal selection child for a tree view.

This is the same idea as 2746b8a329971175f06528e9d623b5246955dc0a except
that the content is completely different, because of course it is.

Related: rhbz#1168950