Rearrange filter.glade
Move the scrolled window on each filter page to be a parent of the
TreeView instead of the parent of the whole page. This way the placement
of the search widgets does not depend on the width of the tree view, and
the TreeView can handle scrolling better without the intermediate
viewport. Replace the grid on each page with nested boxes so that the
various expansion rules make more sense.

Inspired by 1f6f3fb3a31fe23916c8cd29a691b96bc82e5d8a

Related: rhbz#1168950