Make dialogs behave better with timed input validation
Add some functionality to GUIDialogInputCheckHandler to handle setting
button sensitivity, and remove the abstractmethod decorator from
set_status so that classes using GUIDialogInputCheckHandler can use the
new, improved set_status as-is without implementing their own.

Rearrange the running and destroying of GtkDialogs with input validation
to handle the possibility of the dialog returning a response while input
is valid. Move all state saving out of button clicked handlers and into
the return from gtk_dialog_run, ensure that the action buttons return
different response for OK and cancel, and re-run gtk_dialog_run if the
dialog returned a response while input is invalid.