Add simple_replace config file function
When modifying values in config files we want to preserve the rest of
the file as is. SimpleConfig does this for shell-like files using
upper-case KEYS, but other files (eg. INI style) don't use the same
format.

simple_replace takes a list of tuples (key, string) and replaces every
line in the file that starts with the key with the string. It is case
sensitive and the string must be the whole line including key.

eg.

keys=[("switch_1", "switch_1=On")]
simple_replace("/etc/control.conf", keys)

By default if a key isn't replaced in the existing file its string will
be appended to the end of the file. Set add=False to disable this behavior.

(cherry picked from commit f5acc57fda7994ec8168d2027adfa2049bdffbf6)