Depsolve kickstarted packages on the summary hub (#961280)
Pacakage selections with no environments, which kickstarts often are,
were not skipping the depsolve steps, so stop doing that. This way
package errors are displayed before disk formatting takes place.

Change the software spoke slightly to improve its behavior when
interactive modifications are made to a kickstart install. Add a hidden
radio button to allow the environment selection to appear unselected if
there is no environment selected in a kickstart, and switch everything
to the interactive logic if any modifications are made to a kickstart's
environment and group selections.