anaconda-dracut: Mount /dev/mapper/live-rw (#1232411)
systemd added handling of some root=live: cmdline args, dracut adjusted
and started using a systemd generator that is run early, based on the
cmdline, and stopped creating a mount hook when dmsquash-live-root is
called.

There should be a cleaner way for us to interact with dracut, but isn't.
So for now (in the interest of being able to boot and test things) we
will write the dracut mount hook ourselves.