Fix inconsistencies in the payload messages.
The software and source spokes were updating their messages on different
payload events, resulting in some confusing message mismatches.

On STATE_PACKAGE_MD, the statuses displayed would be:

  | Probing storage...              | Downloading package metadata...  |

Followed by STATE_GROUP_MD, which updated to:

  | Downloading package metadata... | Downloading group metadata...    |

Change the source spoke to set the package metadata on the right state
and update the status on STATE_GROUP_MD to match the software spoke.
Moved all the actual messages out of the spokes and into constants.py.