Add keyboard layouts on the row-activated signal.
Besides removing the need for a pile of Gdk constants and flags to
detect double clicks, row-activated is fired on a wider variety of
events, such as hitting Enter when a row is selected.

Also remove the check for whether the button is sensitive, since the
button will take care of that itself.