Add kickstart test for RAID1
This adds a test that creates a 2 disk RAID1 and confirms that /, swap
and /boot are properly mounted. The ks uses numbered md devices to make
checking /proc/mounts easier.

The results checking depends on virt-cat ordering the /dev/mdXXX the
same every time. So far that has proven true.