Add support for autostep and --autoscreenshot (#1234896)
Add the missing support for the autostep [--autoscreenshot]
kickstart command to Anaconda. If just autostep is specified, Anaconda will
iterate over all active spokes just before leaving a hub.
If the --autoscreenshot option specified, it will also take a screenshot of
all hubs and of each spoke.

Also turn the take-screenshot code into a proper function that can
be called when auto-screenshotting and remove the old copy-screenshot
post-script (the last batch of screenshots is taken after the
postscripts have run, so screenshot copying is now handled
elsewhere).

Resolves: rhbz#1234896