dracut: fix boot with inst.ks and no inst.{repo,stage2} (#1238987)
Okay so: commit 4883b96 moved fetch-kickstart-net.sh from the online
hook into initqueue (or initqueue/settled) to fix inst.ks.sendmac.

So when you boot with inst.ks=[URL], the order goes:

1. Network comes up
2. Run `online` hook
   * Schedule kickstart fetch in initqueue
   * Other online tasks from boot args (e.g. updates)
3. Run `initqueue`
   * fetch + run kickstart
     * add `anaconda-netroot.sh` (etc.) to `online` hook
     * `udevadm trigger` disk devices
     * `udevadm trigger` net devices

The problem is, triggering network devices is *not* sufficient to re-run
the `online` hook - this only happens when the NIC actually gets
_configured_, not when the device is triggered.

The fix is pretty straightforward: re-run the `online` hook for active
NICs after we run the kickstart, in order to pick up any new tasks that
the kickstart might have scheduled.

Resolves: RHBZ#1238987