Enable anaconda to use the new cdl ks option, and detect ldl dasds. (#1170656, #1170653)
This checks for the clearpart --cdl ks option and takes appropriate action
if it's specified (automatically format any detected LDL DASDs).

This also adds a general check (during interactive installations) for LDL
DASDs and shows the user the dasdfmt dialog box.

This also bumps the blivet version Requires since is_ldl_dasd was added
to blivet.devicelibs.dasd and is used here.

And to anyone wondering why we show the dasdfmt dialog twice (once for
unformatted, once for LDL), it is to give users the option of formatting
unformatted DASDs but not LDL (or vice versa).

Resolves: rhbz#1170656
Resolves: rhbz#1170653