dracut: fix boot failure waiting for finished/dd.sh
parse-anaconda-dd.sh was creating an (empty) /tmp/dd_todo if there was
nothing to do, but that tricked driver-updates-genrules.sh into thinking
there *was* something to do, so we time out waiting for the dd stuff to
finish.

This patch fixes that by ensuring the file is only created if there is
actually one or more tasks to complete.

Related: rhbz#1084197