Add a helper for building Sphinx docs using mock.
Run ./scripts/make-sphinx-docs and after a while you will have new html
docs in ./docs/html.new/ and instructions on how to update the gh-pages.

This needs to be run from a host with a GUI, otherwise you get Gtk
errors and a coredump from sphinx trying to import the project inside
the mock chroot.