Fix some bad uses of chmod.
When using permission bits to restrict access to the contents of a file,
the permission bits must be set when the file is created, otherwise,
depending on the current umask, there may be a period where any user can
open the file, and the contents will be readable as long as this file
descriptor is open.

Move setting permission bits to the calls to open().