Fix the use of a temporary file in SimpleConfig.write
shutil.move is only an atomic rename if the files are on the same
filesystem, and since /tmp is usually a tmpfs, the temporary file and
the config file were usually not on the same filesystem. Create the
temporary file as a hidden file in the target directory and use
os.rename to move it into place.