Only import readline if readline is necessary.
Weird things can happen when attempting to use readline when not
connected to a tty. Sometimes strange escape characters will appear, or
sometimes python will attempt to manipulate settings that do not exist,
or sometimes libedit will crash the party and break all your glassware.
It is all very mysterious and crash-prone.