Display verbose packaging errors to the user
Sometimes packaging has more information about an error. These are
logged, but not all users know how to read the logs. This adds a
clickable info bar to the source spoke and storage for errors in the
payload.verbose_errors list.

Now when there is a problem fetching .treeinfo or fetching metadata the
exact error will be visible when the info bar is clicked.