act on the right objects when stripping URL protocols (#1243962)
efa549e introduced a _removeUrlPrefix() method which expected
its 'editable' parameter to be a GtkEntry (either the base repo
URL entry box, or an additional repo's URL entry box). However,
it had the on_repoUrl_changed() handler call _removeUrlPrefix()
with its own 'editable' parameter, and then had the protocol
combo box call the on_repoUrl_changed() handler when it was
changed, as well as the URL entry box. The intent is that if,
say, you paste in 'https://www.repo.com' while the protocol is
set to http://, then change the protocol to https://, the
protocol will get stripped - but because 'editable' in this
case is the GtkComboBoxText, not the GtkEntry, it all goes
wrong and crashes. The crash is reproducible by adding a
supplementary repo, entering a URL, then changing the
protocol.

This tweaks the implementation so the repo protocol combo box
handler emits a signal that the repo URL entry box has changed,
instead of the widget using the same handler directly. It also
does the same thing for the base URL protocol handler, which
was missed in the initial commit, so that if you for instance
set the protocol to http://, typed 'https://www.foo.com' into
the URL entry, then changed the protocol to https://, it would
not be stripped. Now that works properly too.