Enable anaconda to use the new rescue mode. (#965985)
This enables the installer to use the new rescue mode, and adds the new
rescue mode file, which now uses the current TUI back-end.

There are a couple of small hacks in here, which were unfortunately
necessary; otherwise it would mean basically rewriting the text mode
code for input handling, which is not prudent at this moment.

Resolves: rhbz#965985