Remove the last usage of newt and get rid of it as a dependency (#965985)
With the new rescue mode it's not really worth it to keep using snack/newt to
let users know they cannot continue because of not enough RAM being availaible
and requiring. We can use simple prints and get rid of the python-newt
dependency.

Thanks to Vratislav Podzimek <vpodzime@redhat.com> for the patch.

Signed-off-by: Vratislav Podzimek <vpodzime@redhat.com>