Wait forever for kickstarts on CDROM (#1168902)
Commit 5d016ef added support for doing cdrom-swapping while doing
"inst.ks=cdrom[:...]". But it turns out that if you're slow to swap the
discs, we'll hit dracut's retry timeout and it'll drop to the emergency
shell.

So, in response to User Requests, we now disable the timeout entirely
when using "inst.ks=cdrom[:...]"; dracut will wait forever for the
kickstart and/or installer media, and if anything goes wrong you're just
stuck and you'll need to reset.

Resolves: rhbz#1168902