Fix Initial PPC PReP Boot Selector Name (#1172755)
The auto partition tool added the PPC PReP Boot partition and named it
prepboot in the selector view in the GUI installer. This was due to the
mountpoint being named by the device format type. Changed the mountpoint
namer so that it is named after the device name instead.

Change required adding BIOS Boot as a name in the mountpoint types to
avoid unexpected behavior as a side effect from this change.

Resolves: rhbz#1172755