Clean up the temporary pools virt-install makes.
For a directory path like /var/tmp/kstest-whatever/disks, virt-install is
making two storage pools:  kstest-whatever, and disks-XX.  It also doesn't
get rid of these pools after it shuts down, and you will eventually end up
with a giant pile of them clogging everything up.

So there are two fixes here to do the clean up:

(1) We know the name of the first pool associated with a single test.  It's
the temporary directory being created in /var/tmp.  So just use the virsh
commands to destroy that pool.

(2) We don't know (and can't figure out) the name of the disks-XX pool.
However, that one is only being created because of the subdirectory that disk
images are being stored in.  So instead store them in /var/tmp/kstest-whatever
and do not create the disks subdirectory any more.