Defer to Fedora distro-wide settings for password strength (#1250746)
Remove the minlen and minquality settings from the default pwpolicy,
and instead pass the buck to libpwquality to determine what is and is
not an acceptable password. Remove strict to allow the user to accept a
bad password.

(cherry picked from commit 92a8c99edd71fa4fc07fdf63e0380f662620e94f)