Get rid of libuser. (#1255066)
The changes made to libuser for python3 support have made it difficult
to work with since it must now be imported from a UTF-8 locale, working
with it in the first place already involved a bunch of gross hacks, and
there's no reason why adding users and groups should need this
complicated of a library. Dump it and use the moderately less quirky
command-line tools.