Do a better job reporting failures from kickstart_tests.
When the failure message is a line from anaconda's logs, it starts with
a timestamp.  Those include colons, so because I am splitting on colons
we get an unhelpful explanation like "03".  Thus, just grab the first
bunch of characters from the log line and display that instead.  It
won't give the complete reason (which could be very long) but it is a
whole lot more helpful than just a number.