Start hubs with the buttons insensitive.
Just because you can't see the buttons doesn't mean you can't click them
(because we attached keyboard shortcuts to them). With the progress hub,
this meant that hitting F12 would reboot in the middle of installing
packages, and in the case of the summary hub there was a potential race
between the hub being created and the button states being initialized.

Start the hubs with the all of the buttons insensitive, and let the code
that was mostly already there update the sensitivities as the status of
the hub changes.