Do not replace the standard streams if not necessary.
Replacing stdout in the middle of the program is about as good of an
idea as it looks like, turns out. While it seems to work fine in text
mode, prints to stdout will just silently disappear after setup_locale
is called by the GUI welcome spoke.

Most of the time replacing the standard streams is not necessary.
Anaconda will have either already started in a UTF-8 locale, or the
first, early call to setup_locale in /usr/sbin/anaconda will switch to a
UTF-8 locale. So for the cases where we do not need to replace the
streams, don't.

(cherry picked from commit 465425ff88ba0d842d1374627452130e34fa3774)