Fix a race between a window continuing and the next starting (#1004477)
Disable the continue button while exiting a standalone so that it cannot
be clicked again while waiting for the next standalone to start. This
was particularly an issue between the welcome spoke and the summary hub,
since the summary hub takes a couple seconds to come up.

Related to this, hide the betanag dialog instead of destroying it, in
case there is still a gap where the welcome spoke can be exited twice.
This way the dialog will simply appear twice instead of being broken and
crashing the second time.

(cherry picked from commit 468dede1f394e7694479f169c02ab2e5f67792ec)