Run nm-connection-editor with the --keep-above flag (#1231856)
It was previously possible to hide the connection editor
window under the Anaconda one by clicking outside of it,
effectively loosing all unsaved data as there is no way to bring
the connection editor window back to foreground once it gets hidden.

So run the nm-connection-editor binary with the --keep-above flag,
which makes sure its window is always above the Anaconda one.

Resolves rhbz#1231856