libnm in spoke: add missing connection for eth device with Configure
On clicking [Configure], default connection to be edited is created,
so next [Configure] or [ON/OFF] will work.

Good enough for the time being, given the cornerity of the case. Running nm-c-e
right on first [Configure] would need some clever event handling (waiting for
connection adding to be finished and reference added to our list.