LVM on RAID kickstart test
This adds a single kickstart test for installation to LVM on top of MD
RAID. Being one of the most complex storage configurations we support the LVM on
RAID definitely deserves testing. However, since such installation is quite a
bit resource-hungry, it would be great if we could get along with this single
test as running multiple such tests could be a problem for our infrastructure
running kickstart tests. That's why this tests the harder case with one of the
LVs requested to be grown to a maximum possible size (which tests blivet's
calculations of free space in a VG using MD RAID PVs and calculations of MD RAID
metadata size).