s/$releasever/rawhide/
We're now using the boot.iso images generated by Fedora instead of ones
generated by us.  This is good because now we're testing something that
is going to end up being the real release, instead of whatever thing I
happened to be making.  However, these images think $releasever is 24
and that does not yet exist.  Thus we need to use rawhide (at least for
now).